Separate class files created for every shape
