+++ Mensuration/bin/.gitignore
@@ -1,1 +1,1 @@
-/Mensuration.class
+/Mensuration/
