--- Mensuration/src/Mensuration.java
@@ -1,749 +1,0 @@
-
-import java.util.Scanner;
-import javax.swing.JOptionPane;
-
-	class Circle{
-	
-		
-		private double radiusCircle;
-		private double areaCircle;
-		private double circumferenceCircle;
-		
-		public double getRadiusCircle() {
-			return radiusCircle;
-		}
-		public void setRadiusCircle(double radiusCircle) {
-			this.radiusCircle = radiusCircle;
-		}
-		public double getAreaCircle() {
-			return areaCircle;
-		}
-		public void setAreaCircle(double areaCircle) {
-			this.areaCircle = areaCircle;
-		}
-		public double getCircumferenceCircle() {
-			return circumferenceCircle;
-		}
-		public void setCircumferenceCircle(double circumferenceCircle) {
-			this.circumferenceCircle = circumferenceCircle;
-		}
-		
-		
-		}
-	
-	class Square{
-	
-		private double sideSquare;
-		private double areaSquare;
-		private double perimeterSquare;
-		
-		
-		public double getSideSquare() {
-			return sideSquare;
-		}
-		public void setSideSquare(double sideSquare) {
-			this.sideSquare = sideSquare;
-		}
-		public double getAreaSquare() {
-			return areaSquare;
-		}
-		public void setAreaSquare(double areaSquare) {
-			this.areaSquare = areaSquare;
-		}
-		public double getPerimeterSquare() {
-			return perimeterSquare;
-		}
-		public void setPerimeterSquare(double perimeterSquare) {
-			this.perimeterSquare = perimeterSquare;
-		}
-
-	
-		
-		}
-
-	class Rectangle{
-	
-		private double lengthRectangle;
-		private double breadthRectangle;
-	
-		private double areaRectangle;
-		private double perimeterRectangle;
-		
-		
-		public double getLengthRectangle() {
-			return lengthRectangle;
-		}
-		public void setLengthRectangle(double lengthRectangle) {
-			this.lengthRectangle = lengthRectangle;
-		}
-		public double getBreadthRectangle() {
-			return breadthRectangle;
-		}
-		public void setBreadthRectangle(double breadthRectangle) {
-			this.breadthRectangle = breadthRectangle;
-		}
-		public double getAreaRectangle() {
-			return areaRectangle;
-		}
-		public void setAreaRectangle(double areaRectangle) {
-			this.areaRectangle = areaRectangle;
-		}
-		public double getPerimeterRectangle() {
-			return perimeterRectangle;
-		}
-		public void setPerimeterRectangle(double perimeterRectangle) {
-			this.perimeterRectangle = perimeterRectangle;
-		}
-		
-		
-	
-	}
-
-	class Triangle{
-	
-		private double baseTriangle;
-		private double heighTriangle;
-		private double sideOne;
-		private double sideTwo;
-		private double sideThree;
-	
-		private double areaTriangle;
-		private double permeterTriangle;
-		
-		
-		
-		public double getBaseTriangle() {
-			return baseTriangle;
-		}
-		public void setBaseTriangle(double baseTriangle) {
-			this.baseTriangle = baseTriangle;
-		}
-		public double getHeighTriangle() {
-			return heighTriangle;
-		}
-		public void setHeighTriangle(double heighTriangle) {
-			this.heighTriangle = heighTriangle;
-		}
-		public double getSideOne() {
-			return sideOne;
-		}
-		public void setSideOne(double sideOne) {
-			this.sideOne = sideOne;
-		}
-		public double getSideTwo() {
-			return sideTwo;
-		}
-		public void setSideTwo(double sideTwo) {
-			this.sideTwo = sideTwo;
-		}
-		public double getSideThree() {
-			return sideThree;
-		}
-		public void setSideThree(double sideThree) {
-			this.sideThree = sideThree;
-		}
-		public double getAreaTriangle() {
-			return areaTriangle;
-		}
-		public void setAreaTriangle(double areaTriangle) {
-			this.areaTriangle = areaTriangle;
-		}
-		public double getPermeterTriangle() {
-			return permeterTriangle;
-		}
-		public void setPermeterTriangle(double permeterTriangle) {
-			this.permeterTriangle = permeterTriangle;
-		}
-		
-		
-		
-	
-	}
-
-	class Parallelogram{
-	
-		private double baseParallelogram;
-		private double heightParallelogram;
-	
-		private double areaParallelogram;
-		private double perimeterParallelogram;
-		
-		
-		
-		public double getBaseParallelogram() {
-			return baseParallelogram;
-		}
-		public void setBaseParallelogram(double baseParallelogram) {
-			this.baseParallelogram = baseParallelogram;
-		}
-		public double getHeightParallelogram() {
-			return heightParallelogram;
-		}
-		public void setHeightParallelogram(double heightParallelogram) {
-			this.heightParallelogram = heightParallelogram;
-		}
-		public double getAreaParallelogram() {
-			return areaParallelogram;
-		}
-		public void setAreaParallelogram(double areaParallelogram) {
-			this.areaParallelogram = areaParallelogram;
-		}
-		public double getPerimeterParallelogram() {
-			return perimeterParallelogram;
-		}
-		public void setPerimeterParallelogram(double perimeterParallelogram) {
-			this.perimeterParallelogram = perimeterParallelogram;
-		}
-		
-		
-	
-	}
-
-	class SemiCircle {
-	
-		private double radiusSemiCircle;
-		private double areaSemiCircle;
-	
-		private double circumferenceSemiCircle;
-
-		public double getRadiusSemiCircle() {
-			return radiusSemiCircle;
-		}
-
-		public void setRadiusSemiCircle(double radiusSemiCircle) {
-			this.radiusSemiCircle = radiusSemiCircle;
-		}
-
-		public double getAreaSemiCircle() {
-			return areaSemiCircle;
-		}
-
-		public void setAreaSemiCircle(double areaSemiCircle) {
-			this.areaSemiCircle = areaSemiCircle;
-		}
-
-		public double getCircumferenceSemiCircle() {
-			return circumferenceSemiCircle;
-		}
-
-		public void setCircumferenceSemiCircle(double circumferenceSemiCircle) {
-			this.circumferenceSemiCircle = circumferenceSemiCircle;
-		}
-		
-		
-		
-	
-	}
-
-	class Rhombus{
-	
-		private double sideRhombus;
-		private double heightRhombus;
-	
-		private double areaRhombus;
-		private double perimeterRhombus;
-		
-		
-		
-		public double getSideRhombus() {
-			return sideRhombus;
-		}
-		public void setSideRhombus(double sideRhombus) {
-			this.sideRhombus = sideRhombus;
-		}
-		public double getHeightRhombus() {
-			return heightRhombus;
-		}
-		public void setHeightRhombus(double heightRhombus) {
-			this.heightRhombus = heightRhombus;
-		}
-		public double getAreaRhombus() {
-			return areaRhombus;
-		}
-		public void setAreaRhombus(double areaRhombus) {
-			this.areaRhombus = areaRhombus;
-		}
-		public double getPerimeterRhombus() {
-			return perimeterRhombus;
-		}
-		public void setPerimeterRhombus(double perimeterRhombus) {
-			this.perimeterRhombus = perimeterRhombus;
-		}
-		
-		
-		
-	
-	}
-
-	class Trapezium{
-	
-		private double sideOneTrapezium;
-		private double sideTwoTrapezium;
-		private double baseOneTrapezium;
-		private double baseTwoTrapezium;
-		private double heightTrapezium;
-	
-		private double areaTrapezium;
-		private double perimeterTrapezium;
-		
-		
-		
-		public double getSideOneTrapezium() {
-			return sideOneTrapezium;
-		}
-		public void setSideOneTrapezium(double sideOneTrapezium) {
-			this.sideOneTrapezium = sideOneTrapezium;
-		}
-		public double getSideTwoTrapezium() {
-			return sideTwoTrapezium;
-		}
-		public void setSideTwoTrapezium(double sideTwoTrapezium) {
-			this.sideTwoTrapezium = sideTwoTrapezium;
-		}
-		public double getBaseOneTrapezium() {
-			return baseOneTrapezium;
-		}
-		public void setBaseOneTrapezium(double baseOneTrapezium) {
-			this.baseOneTrapezium = baseOneTrapezium;
-		}
-		public double getBaseTwoTrapezium() {
-			return baseTwoTrapezium;
-		}
-		public void setBaseTwoTrapezium(double baseTwoTrapezium) {
-			this.baseTwoTrapezium = baseTwoTrapezium;
-		}
-		public double getHeightTrapezium() {
-			return heightTrapezium;
-		}
-		public void setHeightTrapezium(double heightTrapezium) {
-			this.heightTrapezium = heightTrapezium;
-		}
-		public double getAreaTrapezium() {
-			return areaTrapezium;
-		}
-		public void setAreaTrapezium(double areaTrapezium) {
-			this.areaTrapezium = areaTrapezium;
-		}
-		public double getPerimeterTrapezium() {
-			return perimeterTrapezium;
-		}
-		public void setPerimeterTrapezium(double perimeterTrapezium) {
-			this.perimeterTrapezium = perimeterTrapezium;
-		}
-		
-		
-		
-	
-	}
-
-	class Sphere{
-	
-		private double radiusSphere;
-		private double heightSphere;
-	
-		private double areaSphere;
-		private double volumeSphere;
-		
-		
-		
-		public double getRadiusSphere() {
-			return radiusSphere;
-		}
-		public void setRadiusSphere(double radiusSphere) {
-			this.radiusSphere = radiusSphere;
-		}
-		public double getHeightSphere() {
-			return heightSphere;
-		}
-		public void setHeightSphere(double heightSphere) {
-			this.heightSphere = heightSphere;
-		}
-		public double getAreaSphere() {
-			return areaSphere;
-		}
-		public void setAreaSphere(double areaSphere) {
-			this.areaSphere = areaSphere;
-		}
-		public double getVolumeSphere() {
-			return volumeSphere;
-		}
-		public void setVolumeSphere(double volumeSphere) {
-			this.volumeSphere = volumeSphere;
-		}
-		
-		
-		
-	
-	}
-	
-
-	class Cube{
-	
-		private double sideCude;
-	
-		private double surfaceAreaCube;
-		private double volumeCube;
-		
-		
-		public double getSideCude() {
-			return sideCude;
-		}
-		public void setSideCude(double sideCude) {
-			this.sideCude = sideCude;
-		}
-		public double getSurfaceAreaCube() {
-			return surfaceAreaCube;
-		}
-		public void setSurfaceAreaCube(double surfaceAreaCube) {
-			this.surfaceAreaCube = surfaceAreaCube;
-		}
-		public double getVolumeCube() {
-			return volumeCube;
-		}
-		public void setVolumeCube(double volumeCube) {
-			this.volumeCube = volumeCube;
-		}
-		
-		
-		
-	
-	}
-
-	class Cuboid{
-	
-		private double lengthCuboid;
-		private double breadthCuboid;
-	
-		private double volumeCuboid;
-		private double surfaceAreaCuboid;
-		
-		
-		
-		public double getLengthCuboid() {
-			return lengthCuboid;
-		}
-		public void setLengthCuboid(double lengthCuboid) {
-			this.lengthCuboid = lengthCuboid;
-		}
-		public double getBreadthCuboid() {
-			return breadthCuboid;
-		}
-		public void setBreadthCuboid(double breadthCuboid) {
-			this.breadthCuboid = breadthCuboid;
-		}
-		public double getVolumeCuboid() {
-			return volumeCuboid;
-		}
-		public void setVolumeCuboid(double volumeCuboid) {
-			this.volumeCuboid = volumeCuboid;
-		}
-		public double getSurfaceAreaCuboid() {
-			return surfaceAreaCuboid;
-		}
-		public void setSurfaceAreaCuboid(double surfaceAreaCuboid) {
-			this.surfaceAreaCuboid = surfaceAreaCuboid;
-		}
-		
-		
-		
-	
-	}
-
-	class Cone{
-	
-		private double radiusCone;
-		private double heightCone;
-	
-		private double surfaceAreaCone;
-		private double volumeCone;
-		
-		
-		
-		public double getRadiusCone() {
-			return radiusCone;
-		}
-		public void setRadiusCone(double radiusCone) {
-			this.radiusCone = radiusCone;
-		}
-		public double getHeightCone() {
-			return heightCone;
-		}
-		public void setHeightCone(double heightCone) {
-			this.heightCone = heightCone;
-		}
-		public double getSurfaceAreaCone() {
-			return surfaceAreaCone;
-		}
-		public void setSurfaceAreaCone(double surfaceAreaCone) {
-			this.surfaceAreaCone = surfaceAreaCone;
-		}
-		public double getVolumeCone() {
-			return volumeCone;
-		}
-		public void setVolumeCone(double volumeCone) {
-			this.volumeCone = volumeCone;
-		}
-		
-		
-		
-	
-	}
-
-	class Cylinder{
-	
-	
-		private double radiusCylinder;
-		private double heightCylinder;
-	
-		private double surfaceAreaCylinder;
-		private double volumeCylinder;
-		
-		
-		
-		public double getRadiusCylinder() {
-			return radiusCylinder;
-		}
-		public void setRadiusCylinder(double radiusCylinder) {
-			this.radiusCylinder = radiusCylinder;
-		}
-		public double getHeightCylinder() {
-			return heightCylinder;
-		}
-		public void setHeightCylinder(double heightCylinder) {
-			this.heightCylinder = heightCylinder;
-		}
-		public double getSurfaceAreaCylinder() {
-			return surfaceAreaCylinder;
-		}
-		public void setSurfaceAreaCylinder(double surfaceAreaCylinder) {
-			this.surfaceAreaCylinder = surfaceAreaCylinder;
-		}
-		public double getVolumeCylinder() {
-			return volumeCylinder;
-		}
-		public void setVolumeCylinder(double volumeCylinder) {
-			this.volumeCylinder = volumeCylinder;
-		}
-		
-		
-		
-	
-	}
-
-public class Mensuration {
-	
-	Scanner sc = new Scanner(System.in);
-	
-	//Declaring final variable pi
-	static final double PI = 3.1415926536;
-	
-	//Method to display shape menu
-	static int displayMenu(){
-		
-		String choiceStr = JOptionPane.showInputDialog(null,"Select a shape \n 1.Circle \n 2.Square \n 3.Rectangle"
-															+ "\n 4.Triangle \n 5.Parallelogram \n 6.Semi-Circle \n"
-															+ "7.Rhombus \n 8.Trapezium \n 9.Sphere \n 10.Cube \n"
-															+ " 11.Cuboid \n 12.Cone \n 13.Cylinder \n 14.Exit");
-		
-		int choice = Integer.parseInt(choiceStr);
-		
-		return choice;
-		
-	}
-
-	static double findArea(Circle circle) {
-		
-		double radius = circle.getRadiusCircle();
-				
-		double area = PI*radius*radius ;
-		
-		return area;
-	}
-	
-static double findCircumference(Circle circle) {
-	
-		double radius = circle.getRadiusCircle();
-	
-		double circumference = 2*PI*radius;
-		
-		return circumference;
-	}
-	
-	 public static void main(String[] args) {
-		
-		
-		int choice = displayMenu();
-		
-		switch(choice) {
-		
-//Circle
-		case 1:
-			
-			String radiusCircleStr = JOptionPane.showInputDialog(null,"Enter radius in cm");
-			
-			double radiusCircle =  Double.parseDouble(radiusCircleStr);
-			
-			Circle circle = new Circle();
-			circle.setRadiusCircle(radiusCircle); 
-			
-			double area = findArea(circle);
-			double circumference = findCircumference(circle);
-			
-			JOptionPane.showMessageDialog(null,"Area is "+ area+"cm2\nCircumference is "+circumference+"cm");
-			
-			
-			break;
-
-//Square
-		case 2:
-			
-			String sideSquareStr = JOptionPane.showInputDialog(null,"Enter side");
-			
-			double sideSquare =  Double.parseDouble(sideSquareStr);
-			
-			break;
-			
-//Rectangle
-		case 3:
-			
-			String lengthRectangleStr = JOptionPane.showInputDialog(null,"Enter length of rectangle");
-			String breadthRectangleStr = JOptionPane.showInputDialog(null,"Enter breadth of rectangle");
-			
-			double lengthRectangle =  Double.parseDouble(lengthRectangleStr);
-			double breadthRectangle =  Double.parseDouble(breadthRectangleStr);
-			
-			break;
-			
-//Triangle
-		case 4:
-			
-			String baseTriangleStr = JOptionPane.showInputDialog(null,"Enter base of Triangle");
-			String heightTriangleStr = JOptionPane.showInputDialog(null,"Enter heigh of Trianglet");
-			
-			String sideOneTriangleStr = JOptionPane.showInputDialog(null,"Enter side one  of Trianglet");
-			String sideTwoTriangleStr = JOptionPane.showInputDialog(null,"Enter side two of Trianglet");
-			String sideThreeTriangleStr = JOptionPane.showInputDialog(null,"Enter side three of Trianglet");
-			
-			double baseTriangle =  Double.parseDouble(baseTriangleStr);
-			double heightTriangle =  Double.parseDouble(heightTriangleStr);
-			
-			double sideOneTriangle =  Double.parseDouble(sideOneTriangleStr);
-			double sideTwoTriangle =  Double.parseDouble(sideTwoTriangleStr);
-			double sideThreeTriangle =  Double.parseDouble(sideThreeTriangleStr);
-			
-			break;
-			
-//Parallelogram
-		case 5:
-			
-			String baseParallelogramStr = JOptionPane.showInputDialog(null,"Enter base of Parallelogram");
-			String heightParallelogramStr = JOptionPane.showInputDialog(null,"Enter height of Parallelogram");
-			
-			double baseParallelogram =  Double.parseDouble(baseParallelogramStr);
-			double heightParallelogram =  Double.parseDouble(heightParallelogramStr);
-			break;
-			
-//Semi Circle
-		case 6:
-			
-			String radiusSemiStr = JOptionPane.showInputDialog(null,"Enter radius of semi-circle");
-			double radiusSemi =  Double.parseDouble(radiusSemiStr);
-			
-			break;
-			
-//Rhombus
-		case 7:
-			
-			String sideRhombusStr = JOptionPane.showInputDialog(null,"Enter side of Rhombus");
-			String heightRhombusStr = JOptionPane.showInputDialog(null,"Enter height of Rhombus");
-			
-			double sideRhombus =  Double.parseDouble(sideRhombusStr);
-			double heightRhombus =  Double.parseDouble(heightRhombusStr);
-			
-			break;
-			
-//Trapezium
-		case 8:
-			
-			
-			
-			String baseOneTrapStr = JOptionPane.showInputDialog(null,"Enter base one of Trapezium");
-			String baseTwoTrapStr = JOptionPane.showInputDialog(null,"Enter base two of Trapezium");
-			String heightTrapStr = JOptionPane.showInputDialog(null,"Enter height of Trapezium");
-			
-			String sideOneTrapStr = JOptionPane.showInputDialog(null,"Enter side one  of Trapezium");
-			String sideTwoTrapStr = JOptionPane.showInputDialog(null,"Enter side two of Trapezium");
-			
-			double baseOneTrap =  Double.parseDouble(baseOneTrapStr);
-			double baseTwoTrap =  Double.parseDouble(baseTwoTrapStr);
-			double heightTrap =  Double.parseDouble(heightTrapStr);
-			
-			double sideOneTrap =  Double.parseDouble(sideOneTrapStr);
-			double sodeTwoTrap =  Double.parseDouble(sideTwoTrapStr);
-			
-			break;
-			
-//Sphere
-		case 9:
-			
-			String radiusSphereStr = JOptionPane.showInputDialog(null,"Enter radius of sphere");
-			String heightSphereStr = JOptionPane.showInputDialog(null,"Enter height of sphere");
-			
-			double radius =  Double.parseDouble(radiusSphereStr);
-			
-			break;
-			
-//Cube
-		case 10:
-			
-			String sideCubeStr = JOptionPane.showInputDialog(null,"Enter side of cube");
-			double sideCube=  Double.parseDouble(sideCubeStr);
-			
-			break;
-			
-//Cuboid
-		case 11:
-			
-			String lengthCuboidStr = JOptionPane.showInputDialog(null,"Enter length of Cuboid");
-			String breadthCuboidStr = JOptionPane.showInputDialog(null,"Enter breadth of Cuboid");
-			
-			double lengthCuboid =  Double.parseDouble(lengthCuboidStr);
-			double breadthCuboid =  Double.parseDouble(breadthCuboidStr);
-			
-			break;
-			
-//Cone
-		case 12:
-			
-			String radiusConeStr = JOptionPane.showInputDialog(null,"Enter radius of cone ");
-			String heightConeStr = JOptionPane.showInputDialog(null,"Enter height of cone ");
-			
-			double radiusCone =  Double.parseDouble(radiusConeStr);
-			double heightCone =  Double.parseDouble(heightConeStr);
-			
-			break;
-			
-//Cylinder
-		case 13:
-			
-			String radiusCylinderStr = JOptionPane.showInputDialog(null,"Enter radius of Cylinder");
-			String heightCylinderStr = JOptionPane.showInputDialog(null,"Enter height of Cylinder");
-			
-			double radiusCylinder =  Double.parseDouble(radiusCylinderStr);
-			double heightCylinder =  Double.parseDouble(heightCylinderStr);
-			
-			break;
-			
-		case 14:
-			break;
-		}
-		
-			
-		
-		
-	}
-		
-		 
-	}
-
-
