+++ Mensuration/src/Mensuration/Circle.java
@@ -1,0 +1,30 @@
+package Mensuration;
+
+class Circle{
+	
+		
+		private double radiusCircle;
+		private double areaCircle;
+		private double circumferenceCircle;
+		
+		public double getRadiusCircle() {
+			return radiusCircle;
+		}
+		public void setRadiusCircle(double radiusCircle) {
+			this.radiusCircle = radiusCircle;
+		}
+		public double getAreaCircle() {
+			return areaCircle;
+		}
+		public void setAreaCircle(double areaCircle) {
+			this.areaCircle = areaCircle;
+		}
+		public double getCircumferenceCircle() {
+			return circumferenceCircle;
+		}
+		public void setCircumferenceCircle(double circumferenceCircle) {
+			this.circumferenceCircle = circumferenceCircle;
+		}
+		
+		
+		}
