+++ Mensuration/src/Mensuration/Cube.java
@@ -1,0 +1,30 @@
+package Mensuration;
+
+class Cube{
+	
+	private double sideCude;
+
+	private double surfaceAreaCube;
+	private double volumeCube;
+	
+	
+	public double getSideCude() {
+		return sideCude;
+	}
+	public void setSideCude(double sideCude) {
+		this.sideCude = sideCude;
+	}
+	public double getSurfaceAreaCube() {
+		return surfaceAreaCube;
+	}
+	public void setSurfaceAreaCube(double surfaceAreaCube) {
+		this.surfaceAreaCube = surfaceAreaCube;
+	}
+	public double getVolumeCube() {
+		return volumeCube;
+	}
+	public void setVolumeCube(double volumeCube) {
+		this.volumeCube = volumeCube;
+	}
+	
+}
