+++ Mensuration/src/Mensuration/Cuboid.java
@@ -1,0 +1,41 @@
+package Mensuration;
+
+class Cuboid{
+	
+	private double lengthCuboid;
+	private double breadthCuboid;
+
+	private double volumeCuboid;
+	private double surfaceAreaCuboid;
+	
+	
+	
+	public double getLengthCuboid() {
+		return lengthCuboid;
+	}
+	public void setLengthCuboid(double lengthCuboid) {
+		this.lengthCuboid = lengthCuboid;
+	}
+	public double getBreadthCuboid() {
+		return breadthCuboid;
+	}
+	public void setBreadthCuboid(double breadthCuboid) {
+		this.breadthCuboid = breadthCuboid;
+	}
+	public double getVolumeCuboid() {
+		return volumeCuboid;
+	}
+	public void setVolumeCuboid(double volumeCuboid) {
+		this.volumeCuboid = volumeCuboid;
+	}
+	public double getSurfaceAreaCuboid() {
+		return surfaceAreaCuboid;
+	}
+	public void setSurfaceAreaCuboid(double surfaceAreaCuboid) {
+		this.surfaceAreaCuboid = surfaceAreaCuboid;
+	}
+	
+	
+	
+
+}
