+++ Mensuration/src/Mensuration/Parallelogram.java
@@ -1,0 +1,38 @@
+package Mensuration;
+
+class Parallelogram{
+	
+	private double baseParallelogram;
+	private double heightParallelogram;
+
+	private double areaParallelogram;
+	private double perimeterParallelogram;
+	
+	
+	
+	public double getBaseParallelogram() {
+		return baseParallelogram;
+	}
+	public void setBaseParallelogram(double baseParallelogram) {
+		this.baseParallelogram = baseParallelogram;
+	}
+	public double getHeightParallelogram() {
+		return heightParallelogram;
+	}
+	public void setHeightParallelogram(double heightParallelogram) {
+		this.heightParallelogram = heightParallelogram;
+	}
+	public double getAreaParallelogram() {
+		return areaParallelogram;
+	}
+	public void setAreaParallelogram(double areaParallelogram) {
+		this.areaParallelogram = areaParallelogram;
+	}
+	public double getPerimeterParallelogram() {
+		return perimeterParallelogram;
+	}
+	public void setPerimeterParallelogram(double perimeterParallelogram) {
+		this.perimeterParallelogram = perimeterParallelogram;
+	}
+	
+}
