+++ Mensuration/src/Mensuration/Rhombus.java
@@ -1,0 +1,38 @@
+package Mensuration;
+
+class Rhombus{
+	
+	private double sideRhombus;
+	private double heightRhombus;
+
+	private double areaRhombus;
+	private double perimeterRhombus;
+	
+	
+	
+	public double getSideRhombus() {
+		return sideRhombus;
+	}
+	public void setSideRhombus(double sideRhombus) {
+		this.sideRhombus = sideRhombus;
+	}
+	public double getHeightRhombus() {
+		return heightRhombus;
+	}
+	public void setHeightRhombus(double heightRhombus) {
+		this.heightRhombus = heightRhombus;
+	}
+	public double getAreaRhombus() {
+		return areaRhombus;
+	}
+	public void setAreaRhombus(double areaRhombus) {
+		this.areaRhombus = areaRhombus;
+	}
+	public double getPerimeterRhombus() {
+		return perimeterRhombus;
+	}
+	public void setPerimeterRhombus(double perimeterRhombus) {
+		this.perimeterRhombus = perimeterRhombus;
+	}
+	
+}
