+++ Mensuration/src/Mensuration/SemiCircle.java
@@ -1,0 +1,34 @@
+package Mensuration;
+
+class SemiCircle {
+	
+	private double radiusSemiCircle;
+	private double areaSemiCircle;
+
+	private double circumferenceSemiCircle;
+
+	public double getRadiusSemiCircle() {
+		return radiusSemiCircle;
+	}
+
+	public void setRadiusSemiCircle(double radiusSemiCircle) {
+		this.radiusSemiCircle = radiusSemiCircle;
+	}
+
+	public double getAreaSemiCircle() {
+		return areaSemiCircle;
+	}
+
+	public void setAreaSemiCircle(double areaSemiCircle) {
+		this.areaSemiCircle = areaSemiCircle;
+	}
+
+	public double getCircumferenceSemiCircle() {
+		return circumferenceSemiCircle;
+	}
+
+	public void setCircumferenceSemiCircle(double circumferenceSemiCircle) {
+		this.circumferenceSemiCircle = circumferenceSemiCircle;
+	}
+	
+}
