+++ Mensuration/src/Mensuration/Sphere.java
@@ -1,0 +1,38 @@
+package Mensuration;
+
+class Sphere{
+	
+	private double radiusSphere;
+	private double heightSphere;
+
+	private double areaSphere;
+	private double volumeSphere;
+	
+	
+	
+	public double getRadiusSphere() {
+		return radiusSphere;
+	}
+	public void setRadiusSphere(double radiusSphere) {
+		this.radiusSphere = radiusSphere;
+	}
+	public double getHeightSphere() {
+		return heightSphere;
+	}
+	public void setHeightSphere(double heightSphere) {
+		this.heightSphere = heightSphere;
+	}
+	public double getAreaSphere() {
+		return areaSphere;
+	}
+	public void setAreaSphere(double areaSphere) {
+		this.areaSphere = areaSphere;
+	}
+	public double getVolumeSphere() {
+		return volumeSphere;
+	}
+	public void setVolumeSphere(double volumeSphere) {
+		this.volumeSphere = volumeSphere;
+	}
+	
+}
