+++ Mensuration/src/Mensuration/Square.java
@@ -1,0 +1,31 @@
+package Mensuration;
+
+class Square{
+	
+	private double sideSquare;
+	private double areaSquare;
+	private double perimeterSquare;
+	
+	
+	public double getSideSquare() {
+		return sideSquare;
+	}
+	public void setSideSquare(double sideSquare) {
+		this.sideSquare = sideSquare;
+	}
+	public double getAreaSquare() {
+		return areaSquare;
+	}
+	public void setAreaSquare(double areaSquare) {
+		this.areaSquare = areaSquare;
+	}
+	public double getPerimeterSquare() {
+		return perimeterSquare;
+	}
+	public void setPerimeterSquare(double perimeterSquare) {
+		this.perimeterSquare = perimeterSquare;
+	}
+
+
+	
+	}
