+++ Mensuration/src/Mensuration/Trapezium.java
@@ -1,0 +1,59 @@
+package Mensuration;
+
+class Trapezium{
+	
+	private double sideOneTrapezium;
+	private double sideTwoTrapezium;
+	private double baseOneTrapezium;
+	private double baseTwoTrapezium;
+	private double heightTrapezium;
+
+	private double areaTrapezium;
+	private double perimeterTrapezium;
+	
+	
+	
+	public double getSideOneTrapezium() {
+		return sideOneTrapezium;
+	}
+	public void setSideOneTrapezium(double sideOneTrapezium) {
+		this.sideOneTrapezium = sideOneTrapezium;
+	}
+	public double getSideTwoTrapezium() {
+		return sideTwoTrapezium;
+	}
+	public void setSideTwoTrapezium(double sideTwoTrapezium) {
+		this.sideTwoTrapezium = sideTwoTrapezium;
+	}
+	public double getBaseOneTrapezium() {
+		return baseOneTrapezium;
+	}
+	public void setBaseOneTrapezium(double baseOneTrapezium) {
+		this.baseOneTrapezium = baseOneTrapezium;
+	}
+	public double getBaseTwoTrapezium() {
+		return baseTwoTrapezium;
+	}
+	public void setBaseTwoTrapezium(double baseTwoTrapezium) {
+		this.baseTwoTrapezium = baseTwoTrapezium;
+	}
+	public double getHeightTrapezium() {
+		return heightTrapezium;
+	}
+	public void setHeightTrapezium(double heightTrapezium) {
+		this.heightTrapezium = heightTrapezium;
+	}
+	public double getAreaTrapezium() {
+		return areaTrapezium;
+	}
+	public void setAreaTrapezium(double areaTrapezium) {
+		this.areaTrapezium = areaTrapezium;
+	}
+	public double getPerimeterTrapezium() {
+		return perimeterTrapezium;
+	}
+	public void setPerimeterTrapezium(double perimeterTrapezium) {
+		this.perimeterTrapezium = perimeterTrapezium;
+	}
+	
+}
