+++ Mensuration/src/Mensuration/Triangle.java
@@ -1,0 +1,60 @@
+package Mensuration;
+
+class Triangle{
+	
+	private double baseTriangle;
+	private double heighTriangle;
+	private double sideOne;
+	private double sideTwo;
+	private double sideThree;
+
+	private double areaTriangle;
+	private double permeterTriangle;
+	
+	
+	
+	public double getBaseTriangle() {
+		return baseTriangle;
+	}
+	public void setBaseTriangle(double baseTriangle) {
+		this.baseTriangle = baseTriangle;
+	}
+	public double getHeighTriangle() {
+		return heighTriangle;
+	}
+	public void setHeighTriangle(double heighTriangle) {
+		this.heighTriangle = heighTriangle;
+	}
+	public double getSideOne() {
+		return sideOne;
+	}
+	public void setSideOne(double sideOne) {
+		this.sideOne = sideOne;
+	}
+	public double getSideTwo() {
+		return sideTwo;
+	}
+	public void setSideTwo(double sideTwo) {
+		this.sideTwo = sideTwo;
+	}
+	public double getSideThree() {
+		return sideThree;
+	}
+	public void setSideThree(double sideThree) {
+		this.sideThree = sideThree;
+	}
+	public double getAreaTriangle() {
+		return areaTriangle;
+	}
+	public void setAreaTriangle(double areaTriangle) {
+		this.areaTriangle = areaTriangle;
+	}
+	public double getPermeterTriangle() {
+		return permeterTriangle;
+	}
+	public void setPermeterTriangle(double permeterTriangle) {
+		this.permeterTriangle = permeterTriangle;
+	}
+	
+	
+}
